re-modify like button & rearrange component files
re-modify like button & rearrange component files
